Blocks! Blocks! Blocks!
Iterating with blocks means we can avoid creating a lot of arrays. Also,
for isGameOver and its associated methods we can now short-cut: i.e. we
don't have to actually iterate any further as soon as we have found a
move; we then _know_ that it's not game over / loss / draw.